+++ google-cloud-bigquerystorage/src/main/java/com/google/cloud/bigquery/storage/v1/StreamWriter.java
@@ -22,7 +22,9 @@ import com.google.api.gax.core.ExecutorProvider;
 import com.google.api.gax.rpc.TransportChannelProvider;
 import com.google.auto.value.AutoOneOf;
 import com.google.auto.value.AutoValue;
+import com.google.cloud.bigquery.storage.v1.ConnectionWorker.AppendRequestAndResponse;
 import com.google.cloud.bigquery.storage.v1.ConnectionWorker.TableSchemaAndTimestamp;
+import com.google.cloud.bigquery.storage.v1.StreamWriter.SingleConnectionOrConnectionPool.Kind;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import io.grpc.Status;
@@ -32,10 +34,13 @@ import java.io.IOException;
 import java.time.Duration;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.locks.Lock;
 import java.util.logging.Logger;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -76,6 +81,11 @@ public class StreamWriter implements AutoCloseable {
   private final String location;
 
   /*
+   * If user has closed the StreamWriter.
+   */
+  private AtomicBoolean userClosed = new AtomicBoolean(false);
+
+  /*
    * A String that uniquely identifies this writer.
    */
   private final String writerId = UUID.randomUUID().toString();
@@ -98,6 +108,8 @@ public class StreamWriter implements AutoCloseable {
 
   /** Creation timestamp of this streamwriter */
   private final long creationTimestamp;
+
+  private Lock lock;
 
   /** The maximum size of one request. Defined by the API. */
   public static long getApiMaxRequestBytes() {
@@ -139,13 +151,7 @@ public class StreamWriter implements AutoCloseable {
     public ApiFuture<AppendRowsResponse> append(
         StreamWriter streamWriter, ProtoRows protoRows, long offset) {
       if (getKind() == Kind.CONNECTION_WORKER) {
-        return connectionWorker()
-            .append(
-                streamWriter.getStreamName(),
-                streamWriter.getProtoSchema(),
-                protoRows,
-                offset,
-                streamWriter.getMissingValueInterpretationMap());
+        return connectionWorker().append(streamWriter, protoRows, offset);
       } else {
         return connectionWorkerPool().append(streamWriter, protoRows, offset);
       }
@@ -385,6 +391,17 @@ public class StreamWriter implements AutoCloseable {
    * @return the append response wrapped in a future.
    */
   public ApiFuture<AppendRowsResponse> append(ProtoRows rows, long offset) {
+    if (userClosed.get()) {
+      AppendRequestAndResponse requestWrapper =
+          new AppendRequestAndResponse(AppendRowsRequest.newBuilder().build(), this);
+      requestWrapper.appendResult.setException(
+          new Exceptions.StreamWriterClosedException(
+              Status.fromCode(Status.Code.FAILED_PRECONDITION)
+                  .withDescription("User closed StreamWriter"),
+              streamName,
+              getWriterId()));
+      return requestWrapper.appendResult;
+    }
     return this.singleConnectionOrConnectionPool.append(this, rows, offset);
   }
 
@@ -424,11 +441,31 @@ public class StreamWriter implements AutoCloseable {
   public Map<String, AppendRowsRequest.MissingValueInterpretation>
       getMissingValueInterpretationMap() {
     return missingValueInterpretationMap;
+
+  /**
+   * @return if a stream writer can no longer be used for writing. It is due to either the
+   *     StreamWriter is explicitly closed or the underlying connection is broken when connection
+   *     pool is not used. Client should recreate StreamWriter in this case.
+   */
+  public boolean isClosed() {
+    if (singleConnectionOrConnectionPool.getKind() == Kind.CONNECTION_WORKER) {
+      return userClosed.get()
+          || singleConnectionOrConnectionPool.connectionWorker().isConnectionInUnrecoverableState();
+    } else {
+      // With ConnectionPool, we will replace the bad connection automatically.
+      return userClosed.get();
+    }
+  }
+
+  /** @return if user explicitly closed the writer. */
+  public boolean isUserClosed() {
+    return userClosed.get();
   }
 
   /** Close the stream writer. Shut down all resources. */
   @Override
   public void close() {
+    userClosed.set(true);
     singleConnectionOrConnectionPool.close(this);
   }
 
@@ -477,6 +514,15 @@ public class StreamWriter implements AutoCloseable {
   static void cleanUp() {
     testOnlyClientCreatedTimes = 0;
     connectionPoolMap.clear();
+  }
+
+  @VisibleForTesting
+  ConnectionWorkerPool getTestOnlyConnectionWorkerPool() {
+    ConnectionWorkerPool connectionWorkerPool = null;
+    for (Entry<ConnectionPoolKey, ConnectionWorkerPool> entry : connectionPoolMap.entrySet()) {
+      connectionWorkerPool = entry.getValue();
+    }
+    return connectionWorkerPool;
   }
 
   /** A builder of {@link StreamWriter}s. */