+++ google-cloud-bigquerystorage/src/test/java/com/google/cloud/bigquery/storage/v1/ConnectionWorkerTest.java
@@ -81,6 +81,14 @@ public class ConnectionWorkerTest {
         testBigQueryWrite.addResponse(createAppendResponse(i));
       }
       List<ApiFuture<AppendRowsResponse>> futures = new ArrayList<>();
+      StreamWriter sw1 =
+          StreamWriter.newBuilder(TEST_STREAM_1, client)
+              .setWriterSchema(createProtoSchema("foo"))
+              .build();
+      StreamWriter sw2 =
+          StreamWriter.newBuilder(TEST_STREAM_2, client)
+              .setWriterSchema(createProtoSchema("complicate"))
+              .build();
       // We do a pattern of:
       // send to stream1, string1
       // send to stream1, string2
@@ -96,8 +104,7 @@ public class ConnectionWorkerTest {
             futures.add(
                 sendTestMessage(
                     connectionWorker,
-                    TEST_STREAM_1,
-                    createProtoSchema("foo"),
+                    sw1,
                     createFooProtoRows(new String[] {String.valueOf(i)}),
                     i));
             break;
@@ -106,8 +113,7 @@ public class ConnectionWorkerTest {
             futures.add(
                 sendTestMessage(
                     connectionWorker,
-                    TEST_STREAM_2,
-                    createProtoSchema("complicate"),
+                    sw2,
                     createComplicateTypeProtoRows(new String[] {String.valueOf(i)}),
                     i));
             break;
@@ -198,14 +204,19 @@ public class ConnectionWorkerTest {
       // send to stream1, schema3
       // send to stream1, schema1
       // ...
+      StreamWriter sw1 =
+          StreamWriter.newBuilder(TEST_STREAM_1, client).setWriterSchema(schema1).build();
+      StreamWriter sw2 =
+          StreamWriter.newBuilder(TEST_STREAM_1, client).setWriterSchema(schema2).build();
+      StreamWriter sw3 =
+          StreamWriter.newBuilder(TEST_STREAM_1, client).setWriterSchema(schema3).build();
       for (long i = 0; i < appendCount; i++) {
         switch ((int) i % 4) {
           case 0:
             futures.add(
                 sendTestMessage(
                     connectionWorker,
-                    TEST_STREAM_1,
-                    schema1,
+                    sw1,
                     createFooProtoRows(new String[] {String.valueOf(i)}),
                     i));
             break;
@@ -213,8 +224,7 @@ public class ConnectionWorkerTest {
             futures.add(
                 sendTestMessage(
                     connectionWorker,
-                    TEST_STREAM_1,
-                    schema2,
+                    sw2,
                     createFooProtoRows(new String[] {String.valueOf(i)}),
                     i));
             break;
@@ -223,8 +233,7 @@ public class ConnectionWorkerTest {
             futures.add(
                 sendTestMessage(
                     connectionWorker,
-                    TEST_STREAM_1,
-                    schema3,
+                    sw3,
                     createFooProtoRows(new String[] {String.valueOf(i)}),
                     i));
             break;
@@ -248,10 +257,10 @@ public class ConnectionWorkerTest {
         // We will get the request as the pattern of:
         // (writer_stream: t1, schema: schema1)
         // (writer_stream: _, schema: _)
-        // (writer_stream: _, schema: schema3)
-        // (writer_stream: _, schema: _)
-        // (writer_stream: _, schema: schema1)
-        // (writer_stream: _, schema: _)
+        // (writer_stream: t1, schema: schema3)
+        // (writer_stream: t1, schema: _)
+        // (writer_stream: t1, schema: schema1)
+        // (writer_stream: t1, schema: _)
         switch (i % 4) {
           case 0:
             if (i == 0) {
@@ -262,19 +271,23 @@ public class ConnectionWorkerTest {
                 .isEqualTo("foo");
             break;
           case 1:
-            assertThat(serverRequest.getWriteStream()).isEmpty();
+            if (i == 1) {
+              assertThat(serverRequest.getWriteStream()).isEmpty();
+            } else {
+              assertThat(serverRequest.getWriteStream()).isEqualTo(TEST_STREAM_1);
+            }
             // Schema is empty if not at the first request after table switch.
             assertThat(serverRequest.getProtoRows().hasWriterSchema()).isFalse();
             break;
           case 2:
-            assertThat(serverRequest.getWriteStream()).isEmpty();
+            assertThat(serverRequest.getWriteStream()).isEqualTo(TEST_STREAM_1);
             // Schema is populated after table switch.
             assertThat(
                     serverRequest.getProtoRows().getWriterSchema().getProtoDescriptor().getName())
                 .isEqualTo("bar");
             break;
           case 3:
-            assertThat(serverRequest.getWriteStream()).isEmpty();
+            assertThat(serverRequest.getWriteStream()).isEqualTo(TEST_STREAM_1);
             // Schema is empty if not at the first request after table switch.
             assertThat(serverRequest.getProtoRows().hasWriterSchema()).isFalse();
             break;
@@ -290,6 +303,9 @@ public class ConnectionWorkerTest {
 
   @Test
   public void testAppendButInflightQueueFull() throws Exception {
+    ProtoSchema schema1 = createProtoSchema("foo");
+    StreamWriter sw1 =
+        StreamWriter.newBuilder(TEST_STREAM_1, client).setWriterSchema(schema1).build();
     ConnectionWorker connectionWorker =
         new ConnectionWorker(
             TEST_STREAM_1,
@@ -302,7 +318,6 @@ public class ConnectionWorkerTest {
             client.getSettings());
     testBigQueryWrite.setResponseSleep(org.threeten.bp.Duration.ofSeconds(1));
     ConnectionWorker.setMaxInflightQueueWaitTime(500);
-    ProtoSchema schema1 = createProtoSchema("foo");
 
     long appendCount = 6;
     for (int i = 0; i < appendCount; i++) {
@@ -319,11 +334,7 @@ public class ConnectionWorkerTest {
             StatusRuntimeException.class,
             () -> {
               sendTestMessage(
-                  connectionWorker,
-                  TEST_STREAM_1,
-                  schema1,
-                  createFooProtoRows(new String[] {String.valueOf(5)}),
-                  5);
+                  connectionWorker, sw1, createFooProtoRows(new String[] {String.valueOf(5)}), 5);
             });
         long timeDiff = System.currentTimeMillis() - startTime;
         assertEquals(connectionWorker.getLoad().inFlightRequestsCount(), 5);
@@ -331,11 +342,7 @@ public class ConnectionWorkerTest {
       } else {
         futures.add(
             sendTestMessage(
-                connectionWorker,
-                TEST_STREAM_1,
-                schema1,
-                createFooProtoRows(new String[] {String.valueOf(i)}),
-                i));
+                connectionWorker, sw1, createFooProtoRows(new String[] {String.valueOf(i)}), i));
         assertEquals(connectionWorker.getLoad().inFlightRequestsCount(), i + 1);
       }
     }
@@ -393,11 +400,10 @@ public class ConnectionWorkerTest {
 
   private ApiFuture<AppendRowsResponse> sendTestMessage(
       ConnectionWorker connectionWorker,
-      String streamName,
-      ProtoSchema protoSchema,
+      StreamWriter streamWriter,
       ProtoRows protoRows,
       long offset) {
-    return connectionWorker.append(streamName, protoSchema, protoRows, offset, new HashMap());
+    return connectionWorker.append(streamWriter, protoRows, offset);
   }
 
   private ProtoRows createFooProtoRows(String[] messages) {